Extract page getter logic to its own function
