+++ lib/public-api/index.ts
@@ -1,4 +1,5 @@
 export * from './pageFromBlock.class';
 export * from './interfaces';
+export * from './getPageInBoard';
 export * from './mapProperties';
 export * from './validateBody';
